enable GitHub dependabot
this will (partially) automate dependency updates. dependabot will
create PRs if a dependency can be updated automatically.
see the [dependabot documentation][docs] for further information.

as the repository has three actively maintained branches (`main` and
`v1.3` for the plugin and `mixin` for the dashboard mixins). dependabot
by default only targets the default branch (here: `main`) the config has
to to be added twice, the second explicitly for `v1.3`. as `mixin`
doesn't have a gradle build it's only enabled for the GHA updates.

solves #149

[docs]: https://docs.github.com/en/code-security/dependabot/dependabot-version-updates

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>