Removing "leftovers" from the plugin template.
There were some `Rename*` files/classes, this commit is removing them.
However, this step was not straightforward.

TL;DR I had the keep the integration test in `test` folder and disabled
the testingConvention task. I also had to change test testPluginInstalled
a bit because the name of the plugin does no seem to be consistent and
in some cases it equals to a class name (upstream ticket will be opened).

Closes #190

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>